update to latest webpack, babel, react. code cleanup.
