+++ .babelrc
@@ -1,3 +1,7 @@
 {
-  "presets": ["es2015", "react", "stage-1"],
+  "presets": [    "@babel/env",
+    "@babel/preset-react"],
+  "plugins": [
+    "@babel/plugin-proposal-class-properties"
+  ]
 }
