+++ .eslintrc.json
@@ -1,7 +1,7 @@
 {
   "root": true,
   "parser": "babel-eslint",
-  "extends": "airbnb",
+  "extends": ["eslint:recommended", "plugin:react/recommended"],
   "env": {
     "browser": true
   },
