--- components/Loading.jsx
@@ -1,6 +1,0 @@
-import React from 'react';
-
-const Loading = () =>
-  <h1>Loading</h1>;
-
-export default Loading;
