+++ package.json
@@ -3,7 +3,10 @@
   "version": "2.0.0",
   "description": "Single Page Apps for GitHub Pages",
   "scripts": {
-    "start": "webpack -d --watch && live-server --entry-file=404.html"
+    "start": "webpack-dev-server --mode development --hot --history-api-fallback",
+    "build": "webpack --mode production",
+    "preview-http": "http-server",
+    "lint": "eslint src/**"
   },
   "repository": {
     "type": "git",
@@ -12,30 +15,34 @@
   "author": "Rafael Pedicini <code@rafrex.com>",
   "license": "MIT",
   "dependencies": {
-    "axios": "^0.15.3",
-    "current-input": "^1.1.0",
-    "react": "^15.1.0",
-    "react-dom": "^15.1.0",
-    "react-router": "^4.1.1",
-    "react-router-dom": "^4.1.1",
+    "axios": "^0.18.0",
+    "current-input": "^1.1.1",
+    "firebase": "^5.3.0",
     "object.values": "^1.0.4",
-    "prop-types": "^15.5.10",
-    "firebase": "^4.0.0",
-    "paypal-checkout": "^4.0.82"
+    "paypal-checkout": "^4.0.209",
+    "prop-types": "^15.6.2",
+    "react": "^16.4.1",
+    "react-dom": "^16.4.1",
+    "react-router": "^4.3.1",
+    "react-router-dom": "^4.3.1"
   },
   "devDependencies": {
-    "babel-core": "^6.24.1",
-    "babel-eslint": "^7.2.3",
-    "babel-loader": "^7.0.0",
-    "babel-preset-es2015": "^6.24.1",
-    "babel-preset-react": "^6.24.1",
-    "babel-preset-stage-1": "^6.24.1",
-    "eslint": "^3.19.0",
-    "eslint-config-airbnb": "^10.0.0",
-    "eslint-plugin-import": "^1.12.0",
-    "eslint-plugin-jsx-a11y": "^2.0.1",
-    "eslint-plugin-react": "^6.0.0",
-    "live-server": "^1.2.0",
-    "webpack": "^2.6.1"
+    "@babel/core": "^7.0.0-beta.55",
+    "@babel/plugin-proposal-class-properties": "^7.0.0-beta.55",
+    "@babel/preset-env": "^7.0.0-beta.55",
+    "@babel/preset-react": "^7.0.0-beta.55",
+    "babel-eslint": "^8.2.6",
+    "babel-loader": "^8.0.0-beta.4",
+    "html-webpack-plugin": "^3.2.0",
+    "eslint": "^5.2.0",
+    "eslint-plugin-import": "^2.13.0",
+    "eslint-plugin-jsx-a11y": "^6.1.1",
+    "eslint-plugin-react": "^7.10.0",
+    "http-server": "^0.11.1",
+    "optimize-css-assets-webpack-plugin": "^5.0.0",
+    "uglifyjs-webpack-plugin": "^1.2.7",
+    "webpack": "^4.16.3",
+    "webpack-cli": "^3.1.0",
+    "webpack-dev-server": "^3.1.5"
   }
 }
