+++ src/index.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import { render } from 'react-dom';
+import { BrowserRouter as Router } from 'react-router-dom';
+import 'current-input';
+
+import App from './components/App';
+
+render(
+  <Router>
+    <App />
+  </Router>,
+  document.getElementById('root'),
+);
